Update patch set 1
Patch Set 1:

I would like to use the compiler to check that we got every spot along the way. This would involved temporarily changing function names etc to ensure that we changed all of the _ADDR defines to the 7-bit value and any other call sites that passed in a hard coded value without using a define, e.g. [0]

Overall this is the type of change we are looking for.


We can leave ectool alone for now since it is run on the host and uses host command to communicate with the EC. Its command line interface should remain the same for compatibility reasons for now.


[0] https://chromium.git.corp.google.com/chromiumos/platform/ec/+/770f45fac4c7b33e7149ea7906ee5a6eeb852425/board/reef/board.c#370

Patch-set: 1
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>