ish5: fast 32-bit division in scale_us2ticks
The critical section using scale_us2ticks is in __hw_clock_source_set,
which we can optimize as this has 32-bit inupts and outputs.

Make a separate function for this usage, scale_us2ticks_32, to be used
in this critical section.

BUG=b:133190570
BRANCH=none
TEST=ensured scale_us2ticks_32 produced equal values for all us from 0
to 2^16, and really close values (+ or - 4 ticks) for all values from
2^16 to 2^32. Since most inputs to this will be relatively small, this
is sufficent accuracy.

Change-Id: I4f0f928327cb686006c5ebf4ef17becc5e921d9e
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>