WIP: FORTEST: tcs3400: TEST_MODE for generating atime tables
Add a TEST_MODE that is enabled by this CL.  This CL will
enable the alslog mechanism and will enter a test mode where
it will walk the tcs3400 sensor through its various gain and
atime settings, logging the results as it goes so that it can
be post-processed, using the data to populate the range_atime
array used in the anti-saturation adjustment mechanism.

BUG=b:124512628
BRANCH=none
TEST=On an EC console, you will see alslog messages that start
automatically as the test mode begins to walk through the
various gain and atime settings.

Change-Id: I369a122247e25d3dfedd80df4d5d2d9a8710b130
Signed-off-by: Nick Vaccaro <nvaccaro@google.com>