Update patch set 1
Patch Set 1:

Motivation for this is that we have the MPU configured on nocturne_fp. I was going to enable it for hatch_fp, but wanted to see if there's any reason not to just enable it by default for chips that support it since it looks like it theoretically makes things a bit more secure by disabling execution from the non-active flash image.

Patch-set: 1
Reviewer: Gerrit User 1314845 <1314845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>