Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I don't think *all* Cortex-M chips have a MPU. According to the Cortex-M4 TRM, it says that the MPU is optional. It looks like some of the chips that do support a MPU already enable it in their config_chip.h.

Yeah, it does say it's optional. However, it's not enabled for the STM32F12 (what hatch_fp uses), so clearly not all chips that have it have enabled it. If anyone knows of a specific board that doesn't support it, it would be interesting to see how it behaves. Otherwise I guess I have to look through all the datasheets to verify they support it...

Patch-set: 1
Reviewer: Gerrit User 1314845 <1314845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>