Update patch set 17
Patch Set 17:

This CL runs the cable detection before an explicit contract is in place. Planning to change  and run the cable detection once an explicit contract is in place (Planning to upload it in a  separate CL).

The pd flow will look as follows: DFP sends a Discover Identity SOP message. Once ack is received, a Discover command Identity with SOP' packet will be sent. If we receive an ack for SOP', the type of the cable will be stored. Regardless of whether or not the cable responds, the remaining pd flow will proceed.

Patch-set: 17
Reviewer: Gerrit User 1329040 <1329040@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=0