Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> If you're going to be clever about packing into the existing struct, do it this way:
> 
> Use *one* bit in gpio_info.flags to indicate that gpio_info.port is a pointer to the 
ioexpander function table, 

It seems that there is a problem of using 1 bit in gpio_info.flags to indicate if it is an IO-expander IO or not for this API:
  int gpio_get_flags_by_mask(uint32_t port, uint32_t mask)
Parameters passed to this function do not have the enough information to tell what kind of the IO is(i.e. native GPIO or IO-expander IO).

Patch-set: 5
Reviewer: Gerrit User 1132339 <1132339@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>