Update patch set 2
Patch Set 2:

(1 comment)

So, we went through the trouble of adding chargen to the EC and AP, let's just do away with all the complexities of running commands, remembering their output, worrying about sending the next command in time, etc, etc. This seems an overkill to me, let's try getting something simple in which relies on chargen output exclusively.

Also, note that the Cr50 output is flow controlled by the virtue of the fact that if the USB buffer is full we will up the UART FIFO and then stop sending characters, as UART is not ready.

Thus, there is no need to see if there are characters lost on the Cr50 console, its' enough to stress Cr50 by running certain commands which would stimulate some heavy crypto and/or flash activity and verifying that the chargen streams are not corrupted.

Patch-set: 2
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>