Kefka: Implement battery probe cycle
This implements a battery probe cycle in EC autonomously.
Once EC enters the cycle, it wakes up every 2 hours and checks soc
and VBAT.

If both are good, EC sets hibernation timer to 2 hours and sleeps again.
This repeats until a power button is pressed or a battery is cut off
after a battery is drained.

The expected behavior can be verified as follows:
1. Shut down DUT.
2. Unplug AC and leave DUT idle for 60 min.
3. DUT hibernates.
4. DUT wakes up every 2 hours and soc is printed on EC console.
5. DUT cuts off battery when soc goes below 6%.
6. DUT boots by plugging in AC.

Additionally,
- Verify DUT boots when power button is pressed after #3.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/128455667,b/136032811
BRANCH=strago
TEST=See the description above.

Change-Id: I66fd1345e5db1fc2c31bbee59edf36c3b20cab14