config: Add (new) CONFIG_MKBP_WAKEUP_MASK
This commit adds a new CONFIG_* option, CONFIG_MKBP_WAKEUP_MASK.  This
allows a board to specify which MKBP events are allowed to wake the
system when it is in suspend.  Note this CONFIG_* option used to be used
for specifying which host events could wake a system in suspend, but has
since been renamed.

BUG=b:136282898,chromium:786721
BRANCH=None
TEST=With some other code, flash nocturne, suspend DUT, verify that only
the MKBP events in the CONFIG_MKBP_WAKEUP_MASK wake the system up.

Change-Id: Ib4d04418aacab209d0e26703500df119924090b7
Signed-off-by: Aseda Aboagye <aaboagye@google.com>