mkbp_event: Make sure MKBP events can wake the AP.
MKBP was recently refactored to offer choice in the MKBP notification
method.  However, if a board is using a GPIO to notify the AP of a MKBP
event, the MKBP event cannot wake the system up from suspend as is.
This commit simply makes it such that MKBP events can wake the system.
Typically on ARM devices, EC_INT_L is already a wake pin, but on Intel
devices it is not; there's actually a different pin, PCH_WAKE_L which
wakes the system.  By setting that pin as well, we can unify the
behaviour across both ARM and x86 devices.

BUG=b:136272898,chromium:786721
BRANCH=None
TEST=Flash nocturne, suspend DUT, verify that MKBP events can wake the
AP.

Change-Id: If5026bfe3efacbc051f99a180e061c6fd679ce5a
Signed-off-by: Aseda Aboagye <aaboagye@google.com>