Update patch set 5
Patch Set 5:

(2 comments)

This is on the right path to get the compiler to do our checking for us. Thank yo for doing this!

I imagine that you will have a child CL in the same stack that will drop all of the __7b and __8b suffixes that we can submit with this CL, right?

Patch-set: 5
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>