[DCRYPTO Hardening] Adding the ecdsa test that can be invoked through console.
This test's main goal is to be used against future changes in the
dcrypto engine. All it does is comparing signature received from the
ecdsa sign function to the golden results.

Signed-off-by: mschilder@google.com

BRANCH=None

BUG=b:137659935

TEST=something of a test itself and is run by calling the command
through the console connection via host.

The test sequence is following:
- call the dcrypto_ecdsa_sign function on the known input.
- call the dcrypto_ecdsa_verisign on the same input (this function is
added in the following CLs in the chain).
- compare both results to the golden value.

Test passes if both results match to the golden value.

To run the test compile the CR50:
make -j BOARD=cr50 CRYPTO_TEST=1

in the console run:

dcrypto_ecdsa

Change-Id: I07437f6a69ba79bdcce8c92976a374733b17d339