Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> (9 comments)
> 
> I am not sure if configuring PMIC driver in SCP fw a good point.
> Because suspend/resume may happen before the PMIC been configured, and this still leads the the same problem. (I think this could happen in FAFT, please correct me if I am wrong.)
> Could we consider move it to somewhere before SCP fw is loaded?

Hi Yilun/Nicolas,
We can move it to coreboot if the suspend of soc indeed happens before scp fw is loaded.
The only concern is that coreboot cannot identify whether the scp setting is necessary or not for our projects. If it is not a issue for you, I can add it to kukui project initialization.

Patch-set: 6