[WIP] Change the I2C__[78]b addresses to be more specific
Use __7b, __7bf as name extensions for i2c/spi addresses
used in the EC codebase.  __7b indicates a 7bit address by
itself.  __7bf indicates a 7bit address with optional
flags attached.

These extensions are for verification purposes only and
will be removed in the last pass of this ticket.
NOTE: Before submitting this to CQ, make sure to revert
ectool to be in its original state and undo anything
needed to make that happen.

Also removing SMBUS code so I don't have to convert it and
then remove it later

BUG=chromium:971296,chromium:982316
BRANCH=none
TEST=make buildall -j

Change-Id: I2fc3d1b52ce76184492b2aaff3060f486ca45f45
Signed-off-by: Denis Brockus <dbrockus@chromium.org>