wip: ideas for usb-c statemachine
This CL does not build. It is for discussion on the following points:

 - OBJ is renamed to context (CTX) for current and last state
 - State definition now includes its parent (no need for the boiler
plate function that takes in a signal now)
 - The init_state, set_state, and exe_state have been re-written to take
advanages of new state machine definition (seems simpler now, but I may
have missed a use case, please check)
 - Explicitly create the USB states at the bottom of the file with all
of the statics. Dont need to use macros (even though I did suggest them)
  - Uses null when we do_nothing instead of calling into a function
 - Created a "private" enum in the C file that lists all of the states
in the file, that we can use to refer to a state (it is also the
index into the states array for that state).
 - Changed prototype of state function to return void, since we aren't
really using the return value and it cleans up a lot of return 0 that
aren't needed.
 - Add const to int port since we can (and should) for typedef

Change-Id: I057ae3cc7812739dd0098721fd0db57d9477ef4e
Signed-off-by: Jett Rink <jettrink@chromium.org>