Update patch set 17
Patch Set 17:

> Patch Set 17:
> 
> There are still 19 unresolved comments. Which ones are still open?
> 
> Also, while I was reviewing, I had a few ideas that were easier to express in code while I was experimenting. I upload CL:1691108 to suggest a slightly different way to express the state machine. If we like it I can go do the work to make it build of ToT, but I didn't want to waste effort if we changed it while we were discussing it.

Sorry, I answered the comments in patchset 15 but didn't hit the resolved button.

I like your new architecture but this is a major change. And if we are going to make the Hatch deadline, I think we should focus on getting the TypeC/PD functionality correct and then start improving the design. I still have another CL, with the bulk of the PD functionality, that needs to be reviewed after this one.

Patch-set: 17
Reviewer: Gerrit User 1149894 <1149894@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>