Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 6:
> > 
> > > Patch Set 5:
> > > 
> > > (9 comments)
> > > 
> > > I am not sure if configuring PMIC driver in SCP fw a good point.
> > > Because suspend/resume may happen before the PMIC been configured, and this still leads the the same problem. (I think this could happen in FAFT, please correct me if I am wrong.)
> > > Could we consider move it to somewhere before SCP fw is loaded?
> > 
> > Hi Yilun/Nicolas,
> > We can move it to coreboot if the suspend of soc indeed happens before scp fw is loaded.
> > The only concern is that coreboot cannot identify whether the scp setting is necessary or not for our projects. If it is not a issue for you, I can add it to kukui project initialization.
> 
> Hi Yulin/Nicolas,
> Here is another concern:
> Does sco always run in 330Mhz(ULPOSC2)?
> If yes, we don't need pmic driver in scp image.

AFAICT, that's the plan. The only time we thought about lower clock speed is in S3, but it appears difficult as there is no other good clock we can use.

> If not, scp would enter suspend or adjust clock speed, and then it would need pmic driver to adjust core voltage at runtime.

Right. I assume this would be to run the SCP at >330Mhz, right?

I'd say: move this to coreboot for now, and we can revisit later if this becomes a problem?

Thanks.

Patch-set: 7
Reviewer: Gerrit User 1129199 <1129199@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>