Update patch set 2
Patch Set 2: Code-Review-2

I don't think we should be doing this. Touching a whole bunch of files just to change the style of bug format doesn't seem right. If this is just to make it more convenient for your IDE choice, I don't think that's a good enough reason to change all occurrences in the repo. I honestly believes that this belongs in the IDE.

As far as I know, there are no style requirements for bug references in the EC. If we want to add one, we can, but I don't think we should change everything else.

Patch-set: 2
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2