driver/tcs3400: make error message in sensor read less spammy.
In EMULATED_IRQ_EVENT mode, we tries to estimate the time between a read
request and data ready. If the estimated time is shorter, an error
message is printed to the console on every retry.

On Kukui, this caused ec console spamming one line per second.

Change the behavior to print only when returning error.

No need to adjust the estimate time because one retry per read is still
good enough.

BUG=b:137345336
TEST=manually, verify that error message disappeared.
BRANCH=None

Change-Id: Ie85e8f1a80ea36ca5c50f57a19eba35f3f79f86d
Signed-off-by: Ting Shen <phoenixshen@google.com>