util: Unify separate flash_fp_mcu scripts into single common one
For fingerprint firmware we are moving towards a model where the "board"
(in EC terminology) is a specific MCU+FP_SENSOR combination and not tied
to the main system board that it's connected to (e.g., "hatch",
"nocturne", etc.). This change decouples flash_fp_mcu from the EC
"board".

BRANCH=none
BUG=b:136678758,b:137108509
TEST=make buildall -j
     hatch:
       flash_fp_mcu /opt/google/biod/fw/hatch_fp_v2.0.1359-6f54be08d.bin
     nocturne:
       flash_fp_mcu /opt/google/biod/fw/nocturne_fp_v2.0.1765+87bb17a39.bin
     nami:
       flash_fp_mcu /opt/google/biod/fw/nami_fp_v2.2.144-7a08e07eb.bin

Cq-Depend:chromium:1705055
Change-Id: Idfe298f59ab9df8657a570cc47e956b4e94ee1a1
Signed-off-by: Tom Hughes <tomhughes@chromium.org>