Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > Ping. Getting multiple requests per day for this from vendors. Can I get a +2 so I can land it?
> > 
> > Let me know if there's some reason this definitely cannot land and I'll fix it.
> 
> My concern is there is too much temporary & duplicate info in these docs. I expect these docs to use proper pointers to source code or other docs on chromium.org. I certainly don't want to see future CLs which will update these docs too often. We traditionally shared some of the contents via 'care & feeding' doc with vendors. I feel the contents need to be filtered & brushed up to be more lasting.
> 
> If there is urgency, why don't you share the contents via google doc? You probably get feedback which will be useful to update these docs.

The main reason I'm not putting this in a Google Doc is that it doesn't scale. I don't want to spend any of my time repeating myself to vendors or anyone else that needs this information, which I often do with the docs method. The care and feeding docs go out of date really quickly because they're hard to find and not open source. The result is that we're just hiding messy documentation in Google Docs. Over time it becomes really hard to disentangle what aspects we could put into public documentation and what is outdated so it's really hard to ever solve the problem. With the fingerprint docs I'm trying to make everything opensource from the start to avoid that problem. It may be a little messy or incomplete along the way, but that's ok. Documentation should be free to update (i.e., I should be able to update fingerprint-specific docs three to four times a day with no approval needed -- not that I would be doing this) since it shouldn't impact anyone else's code.

> Maybe there is a better place somewhere under http://www.chromium.org/chromium-os?

No, the documentation on the wiki is being migrated to Markdown over time, so we don't want to add more things to the outdated sites page. Markdown is way easier to deal with programmatically.

https://groups.google.com/a/google.com/d/msg/chromeos-chatty-eng/sICu82J2ejQ/8vE0wuqRCQAJ

https://groups.google.com/a/google.com/d/msg/chromeos-chatty-eng/sICu82J2ejQ/5bqLThCbCQAJ

Patch-set: 5
Reviewer: Gerrit User 1314845 <1314845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>