Remove %T printf specifier
In order to enable compile-time printf format checking, non-standard
printf format specifiers should be removed. %T is not supported generally,
and had the awkward additional property that it didn't take any arguments.
Move the timestamp printing support out into an explicit function call.

For the most part, there should be no observable functional change here,
with the following exceptions:
 * board/zinger/usb_pd_policy.c: Remove timestamp from pd_custom_vdm. This
one was nasty because it is using debug_printf directly, and support for
ccprints isn't compiled in. If the timestamp in these prints is deemed
important, I could hoist a copy of the timestamp print function out near
the debug_printf definitions. I made a judgment call that just removing
the timestamp was better, feel free to shout if it's not.
 * drivers/accel_lis2dh.c: Prints in is_data_ready, read, and init now have
newlines at the end.
 * drivers/accelgyro_lsm6ds0.c: Prints in is_data_ready and read now have
newlines at the end.
 * drivers/accelgyro_lsm6dsm.c: Print in is_data_ready now has a newline
 * drivers/gyro_l3gd20h.c: Prints in is_data_ready, read now have a
newline.

BUG=chromium:984041
TEST=make -j buildall, also basic sanity test on Hatch
BRANCH=None

Cq-Depend:chrome-internal:1473305
Change-Id: I83e45b55a95ea27256dc147544ae3f7e39acc5dd
Signed-off-by: Evan Green <evgreen@chromium.org>