WIP: ps8xxx: stub out enter_low_power_mode
the ps8751 really does not work well with explicit low power mode; it
enters low power mode on its own in about 2 seconds.

when the ps8751 is awake, it will take the low power command, but also
generate an alert.  when we try to read the alert register, it fails
because the chip is in low power mode, so we go through the normal
process of waking it up, including re-initializing the chip.
we then process the alert and power down the chip again.

but this time around the i2c transaction for power down command
fails (we don't check this status).  the reason the i2c transaction failed
is very likely because the chip processed the I2CIDLE command
and didn't complete the i2c transaction.

BRANCH=none
BUG=b:137622553

Change-Id: I01b864791ae27a81a4512ddb752329e59be6a8c6
Signed-off-by: Caveh Jalali <caveh@chromium.org>