ish: snowball linker cleanup fix
The following bug

https://buganizer.corp.google.com/issues/136002955

indicates after my initial move of snowball to use the
linker map instead of hard defined addresses that
0xFF801E80 was reading back as all zeroes.

The change that was made for this is
https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1664593

The first change altered the define names from CONFIG_AON_ROM_*
to be CONFIG_AON_PERSISTENT_*, this could not be the problem.

The second change was to stop using the hard defined addresses
and use the linker to map snowball instead.  I verified the
addresses as being the following:

0xFF801E80 CONFIG_AON_PERSISTENT_BASE
0xFF801F80 snowball
0xFF801FF0 snowball.uma_base_hi
0xFF801FF4 snowball.uma_base_lo
0xFF801FF8 snowball.uma_limit
0xFF801FFC snowball.fw_offset

This is only used in the aon task. These are all consistent
with what the previous defines were.

The only other thing that looks changed was where volatile
was being used.  I will adjust the volatile to be on each
structure field and see if that makes a difference.

Jack, how did you test this? I want to be able to verify
it before it gets merged.  I am not sure how a change in
snowball affects the values in 0xFF801E80, which is 256
bytes before snowball.

BUG=b:136002955
BRANCH=none
TEST=make buildall -j
TEST=get info from Jack to make sure the problem is gone

Change-Id: Ibb6dbd3a4414b5c546e99f5ad7e0409250de6256
Signed-off-by: Denis Brockus <dbrockus@chromium.org>