ish: fix snowball structure linker placement
The following bug

https://buganizer.corp.google.com/issues/136002955

indicates after my initial move of snowball to use the
linker map instead of hard defined addresses that
0xFF801E80 was reading back as all zeroes.

The change that was made for this is
https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1664593

I double checked the map files and everything looked
good.  For some reason using the linker to map the location
of this structure, the system does not run properly.
If I remove the link map placement of this structure
then the issue goes away.  I looked at how aon_share
was placed at the specific address in AON memory and
this CL is doing the same thing and it is working.

I think this is the way we should keep this fix and not
try to get the linker map to place this structure where
we want it.

BUG=b:136002955
BRANCH=none
TEST=make buildall -j
TEST=get info from Jack to make sure the problem is gone

Change-Id: Ibb6dbd3a4414b5c546e99f5ad7e0409250de6256
Signed-off-by: Denis Brockus <dbrockus@chromium.org>