Update patch set 2
Patch Set 2: Code-Review-1

Actually, should we be bringing this into master at all? This seems like it was necessary for a specific board design. If this design pattern was not a one-off, then maybe we should add a CONFIG_* option for it?

Patch-set: 2
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1