Update patch set 2
Patch Set 2: Verified+1 Auto-Submit+1

I backed out the linker portion of this CL and just pointed the structure by pointer assignment to the start of the snowball area. This make this work. The problem looks like it has to do with linking the snowball portion into higher memory but before the default .data section is declared.  If I move this to be after the .data section linking then snowball does not land in the .data.snowball section but hits the default instead.  Let's push this so it works and we can rethink the linking option later.

Patch-set: 2
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1
Label: Auto-Submit=+1