driver: lis2dwl: Add support for lis2dwl
Though lis2dwl has a similar name as lis2dh/lis2de, it has
different register layout and field definitions. So it does
not use the same code as lid2dh and this makes code cleaner.

Since the sensor is currently only used as LID sensor, FIFO
and interrupt are not supported.

BUG=b:137945787, b:138879565
BRANCH=none
TEST=boot Akemi board, use 'accelinfo on' to watch lid accel x/y/z
     output and make sure the datas are correct.

Change-Id: I18da4ebe14117620f459f5354d55c2d9ae876d8b
Signed-off-by: Paul Ma <magf@bitland.corp-partner.google.com>