Update patch set 2
Patch Set 2: Code-Review+1

Previously this was used on grunt when I was testing ec codec on grunt.
Grunt uses feature to determine whether it needs to create a device for audio in mfd device creation flow.
I think we will need it for future AMD projects where ec codec is needed.
But I would also agree that we can add that in the future when we need it so the code is clean.

Patch-set: 2
Reviewer: Gerrit User 1001485 <1001485@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1