Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> Why does everything now have _flags at the end of it?
> 
> A lot of variables have confusing names now. Take common/i2c_master.c, in scan_bus: addr_flags.
> 
> I read that as... flags of the address, not as addr with flags in it (though I don't know what that even means).
> 
> Is it possible to just remove the suffixes on everything?

It is possible, but I would not advise to do so.  The slave address was used through the system to mean 7-bit, 8-bit and 7-bit with flags.  The flags included the SPI override and big endian. The flags were not always in a consistent location.  Except for locally within a driver that touches the hardware that uses the 8-bit value, all slave addresses are now 7-bit.  There is a difference between 7-bit and 7-bit with flags in that the flags being optional, depending on how the caller is using the interface, the 7-bit without flags is the only way to compare just the slave address.  I am sorry that changing addr to addr_flags is confusing but the amount of confusion that can be reduced by knowing the flags may be set could help in future development. The suffix _flags is the most common usage of a slave address.  If you would prefer this be the case that does not include a suffix, please recommend a naming standard for 7-bit without flags, 7-bit with flags and 8-bit.

Patch-set: 10