mock: Add mock structure for fuzzers and tests
This introduces the common/mock and include/mock directories
which hold mock implementations of other common libraries.

The general idea it to create mocks which can be used in
simple scenarios, by providing meaningful default return
values and functionality, and more complicated scenarios
like fuzzers, where the return values need to continuously change.

The build system has been adapted to allow the inclusion of
a new .mocklist file for tests and fuzzers. This file specifies
exactly which mocks will be pulled into the build for a given
test/fuzz. In order to maintain cleanliness, this file is optional.

Examples:
* http://crrev.com/c/1682945/17 makes use of three different mocks,
  one of which is the rollback mock.
* An upcoming rollback unit test ( http://crrev.com/c/1686460 ),
  needs to pull in mocks that support the rollback interface, but
  explicitly cannot pull in the rollback mock.

BRANCH=none
BUG=b:116065496
TEST=make buildall -j

Change-Id: Ib87b1a93b6d73309afaf7115276ead49218598ff
Signed-off-by: Craig Hesling <hesling@chromium.org>