Update patch set 24
Patch Set 24:

(2 comments)

Can we defer on the refactoring changes to follow up CLs? I think we'd like to merge something that's functionally correct. We can (and should) clean it up, but I think those should be follow up CLs, otherwise the review just drags on for months and months. For example, Jett proposed an alternative state machine framework which may clean things up significantly. It might not be worth spending time on it now, when we may no longer use this state machine implementation. There's also the fact that we want to create a common utilities module to be shared between the two stacks. 

Keep in mind that this is just the Type-C state machine. There's still the power delivery state machine to come.

Patch-set: 24
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>