nautilus : Do not override charge on dead battery
charger_profile_override() change charging voltage & current
as battery temperature. The flag for bad temperature is set
when battery is not responsive then the current is set to 0.
Charge does not charge dead battery and the battery can't wake.
Then the system can't power on with dead battery.

BUG=b:138413964
BRANCH=none
TEST=flash EC and check if dead battery is charged.

Change-Id: I6af5c2d328df1c1be16261d581cdcafeb6036eaa
Signed-off-by: YongBeum Ha <ybha@samsung.com>