Zork: Change host_cmd to be ESPI
This functionality requires a chipset_handle_espi_reset_assert
to be defined.  I created on in power/stoney.c but it is not
filled in for what is needed. Not sure anything needs to be
done but leaving a placeholder to make sure to follow up for
verification.

BUG=none
BRANCH=none
TEST=make buildall -j

Change-Id: I4c21e75c0ed7cafbc7f2058e9bd3c36bc8314c3a
Signed-off-by: Denis Brockus <dbrockus@chromium.org>