chip/host: Add check for TEST_BUILD in trng.c
Due to the possibly unnoticeable security vulnerability
it would cause if the dummy host trng driver made it's way
into production (non-test) code, this change adds a small check
to ensure it is being used only in test or fuzz builds.
This isn't as much of a concern for other dummy host drivers.
For example, including a fake spi driver may simply manifest as
broken feature.

BRANCH=none
BUG=none
TEST=make buildall -j

Change-Id: I31138976566d39ed44d905bbb2c43c5f6decbaf4
Signed-off-by: Craig Hesling <hesling@chromium.org>