Restore showboards target back into the makefile
The showboards target is used to generate a list of supported boards for
various build scripts to use, most notably the coverity builder.  The
new print-boards target is a nice addition for users to see the list of
boards, but a simple list is better for scripting.

showboards was deleted in commit 053491b560

BUG=None
BRANCH=None
TEST='make showboards' gives the list of boards again.

Change-Id: I0064cf28f8f33c8dea367cad4360127a19eedf04
Signed-off-by: Martin Roth <martinroth@chromium.org>