Update patch set 1
Patch Set 1:

I apologize. I made that destructive change after seeing 0 hits for this build target on code search. I should have contacted Martin first.

It should be noted that print-boards maintains the same functionality for
scripting purposes[it checks if a tty is connect, if not it uses a simple newline format similar to `ls|cat`], but I understand if you need the same build target name.

Patch-set: 1
CC: Gerrit User 1334530 <1334530@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>