Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> just out of curiosity - would it yield the same result if all these functions were put in a separate file and were not included when the standard library function is linked in?

Not including the file isn't enough. We need these functions for linking with EC code, but we also need to hide them when linking outside libraries for tests.

Moving them all to a single object file would allow us to potentially sent the visibility for the whole object file and get rid of some of the macros, but the headers are still tricky. Something like HIDE_EC_STDLIB is needed to exclude declarations that overlap with the C standard library to avoid conflicts.

The change Craig is making is nice in that it doesn't negatively affect the current behavior and fixes a problem he ran into when trying to enable code coverage.

Patch-set: 1
Reviewer: Gerrit User 1265562 <1265562@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>