flapjack/tcs3400: use LUX table
Instruct tcs3400 driver to use the LUX tables in the
anti-saturation algorithm.  This allows the driver to
auto-train to about 90% of saturation in a single cycle
instead of many.

BUG=b:124512628
BRANCH=master
TEST=Flash and boot flapjack, verify that ALS and RGB sensors
are still generating data.  I used alslog patch and enabled
ALS logging in EC console via "alslog 2047".  Verify that under
a constant light source, the adjustment mechanism correctly drives
the ALS values such that they land in the sweet spot between 90
to <100% of saturation.

Change-Id: I5ac19ca1bf826fdf8ccc767c61e4773407188e13
Signed-off-by: Nick Vaccaro <nvaccaro@google.com>