Update patch set 23
Patch Set 23:

> Patch Set 22:
> 
> (3 comments)
> 
> Just curious why there's an option to turn off the new algorithm.
> Seems like the new one is much better, is there any advantages to not using TCS_USE_LUX_TABLE?

Using the LUX tables requires that the driver be able to properly calculate the current LUX.  To do that, the cover scale and k_channel scales for the device need to have been calculated and added to the code, or the lux calculation will be off, which could cause the wrong step value to be grabbed from the table.  Once the k_channel_scale and cover_scale values for a board have been defined, and the CONFIG_RGB_XYZ_TRANSLATION enabled (requires lens coefficients to be defined for board, which is a per-board value determined at factory calibration time), then TCS_USE_LUX_TABLE could be defined for best operation.

Patch-set: 23
Reviewer: Gerrit User 1185940 <1185940@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>