WIP: ps8xxx: disable DCI mode
DCI mode is auto-enabled by default.  we don't actually support DCI
(intel specific SoC debug path), so we can explicitly disable DCI
mode.  doing so, saves about 40mW on the 3.3v rail when USB2 devices
or USB-C to USB-A dongles are left plugged in.  this is particulary
relevant in sleep mode as this accounts for a significant portion
of the system power consumption.

BUG=b:119875949
BRANCH=none
TEST=verified power consumption drops using sweetberry, USB devices
	still functional across suspend/resume.

Change-Id: Id13630425c78965d2ac4f2e97715374ae0640d23
Signed-off-by: Caveh Jalali <caveh@chromium.org>