mkbp_event: Only notify MKBP via hostevent in suspend
If a board is using CONFIG_MKBP_USE_GPIO_AND_HOST_EVENT, make sure that
the MKBP event notiification via host event only occurs in suspend.
Since MKBP events are a part of the
HOST_EVENT_ALWAYS_REPORT_DEFAULT_MASK, an MKBP host event could still be
set in S0, but it will not trigger an SCI since the event is not in the
SCI mask.  This would cause the board to prematurely wake up when
suspending due to the lingering event.

BUG=none
BRANCH=none
TEST=Flash nocturne, boot to S0, suspend, verify that no spurious
wakeups occur.

Change-Id: I2e3196e836934d87d1d5efefff26f58c9e2bc3b2
Signed-off-by: Aseda Aboagye <aaboagye@google.com>