g: allow I2CS operate without hardware resets
It is not always possible to rely on PMU for resetting the I2CS
controller. Most of the AP firmware versions deploy the 'I2C unwedge'
cycle when coming out of reset, but not all of them, this is why Cr50
needs to be able to recover on its own in case there was a crash and
the I2C bus was left mid transaction with the H1 holding down the SDA
line.

The solution requires the following steps:

   - change the pin multiplexer settings such that SDA and SCL lines
     of the I2CS controller are connected to two dedicated GPIOs
     instead of the actual bus. Use internal 'virtual' pads for this
     instead of actual GPIO pins.

  - issue several clocks to make sure that the controller finishes
    pushing out the byte currently in the TX buffer

  - using the GPIOs clreate the bus I2C Stop condition to prepare the
    controller to the next transaction

   - restore pinmux settings so that the controller is connected back
     to the external bus

Value read through the SDA field of the I2CS_READVAL register is used
to determine if the bus could potentially be in the wedges state.

Console command 'pinmux' was modified to show state of the virtual
pads used to connect the new GPIOs to the I2CS controller.

BRANCH=cr50, cr50-mp
BUG=b:135772657

TEST=the test was ran on a Pyro device, which uses I2C for
     communication with H1 and which AP firmware does not deploy the
     'I2C unwedge' cycle.

     Test instrumentation involved setting a Chrome OS startup file
     such that once booted, the AP starts continuously polling TPM for
     value of an NVMEM index, creating I2C traffic. The host
     workstation sends the 'apreset cold' command to the EC within a
     few seconds of Chrome OS coming up.

     First run a special Cr50 image which is not resetting I2CS using
     PMU on TPM restarts, is was not trying to unwedge the stuck I2C
     bus. On five experiments, it takes on average 32 reboots for
     until I2C bus is locked up and the DUT falls into recovery.

     Then loaded the Cr50 image with this patch and ran the test
     again, it survived for 150 cycles without a problem.

Change-Id: Iffec33f97557e3acfd1cd5fb76ba158f8c23b608
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>